feat: complete Spring Boot backend with JWT security
- Implement JWT authentication (register/login endpoints)
- Add Spring Security configuration with role-based access
- Secure all controllers with @PreAuthorize annotations
- Configure H2 database for development
- Add comprehensive testing suite
- Fix Java version compatibility issues
- Setup project structure with entities, DTOs, services, repositories

+++ backend/src/main/java/com/mobility/mobility_backend/repository/UserRepository.java
@@ -8,13 +8,13 @@ import org.springframework.stereotype.Repository;
 import com.mobility.mobility_backend.entity.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
-    Optional<User> findByUsername(String username);
+	Optional<User> findByUsername(String username);
 
-    Optional<User> findByEmail(String email);
+	Optional<User> findByEmail(String email);
 
-    boolean existsByUsername(String username);
+	boolean existsByUsername(String username);
 
-    boolean existsByEmail(String email);
+	boolean existsByEmail(String email);
 }
